+++ src/ts/Test/Unit/EventHandlers/VastOverlayEventHandlersTest.ts
@@ -21,7 +21,6 @@ import { VastEndScreen } from 'Views/VastEndScreen';
 import { MetaDataManager } from 'Managers/MetaDataManager';
 import { FocusManager } from 'Managers/FocusManager';
 import { Request } from 'Utilities/Request';
-import { ComScoreTrackingService } from 'Utilities/ComScoreTrackingService';
 import { MOAT } from 'Views/MOAT';
 import { MoatViewabilityService } from 'Utilities/MoatViewabilityService';
 import { OperativeEventManagerFactory } from 'Managers/OperativeEventManagerFactory';
@@ -45,7 +44,6 @@ describe('VastOverlayEventHandlersTest', () => {
     let request: Request;
     let vastAdUnitParameters: IVastAdUnitParameters;
     let vastOverlayEventHandler: VastOverlayEventHandler;
-    let comScoreService: ComScoreTrackingService;
     let moat: MOAT;
     let sandbox: sinon.SinonSandbox;
 
@@ -100,8 +98,6 @@ describe('VastOverlayEventHandlersTest', () => {
             campaign: campaign
         });
 
-        comScoreService = new ComScoreTrackingService(thirdPartyEventManager, nativeBridge, deviceInfo);
-
         vastAdUnitParameters = {
             forceOrientation: Orientation.LANDSCAPE,
             focusManager: focusManager,
@@ -110,7 +106,6 @@ describe('VastOverlayEventHandlersTest', () => {
             clientInfo: clientInfo,
             thirdPartyEventManager: thirdPartyEventManager,
             operativeEventManager: operativeEventManager,
-            comScoreTrackingService: comScoreService,
             placement: TestFixtures.getPlacement(),
             campaign: campaign,
             configuration: configuration,