+++ src/ts/Utilities/CustomFeatures.ts
@@ -7,6 +7,10 @@ import { StorageError } from 'Native/Api/Storage';
 
 export class CustomFeatures {
 
+    public static isIosVideoCachingEnabled(abGroup: number): boolean {
+        return abGroup === 14 || abGroup === 15;
+    }
+
     public static isExampleGameId(gameId: string): boolean {
         return gameId === '14850' || gameId === '14851';
     }
@@ -23,51 +27,4 @@ export class CustomFeatures {
         return new AdUnitStyle({ctaButtonColor: '#167dfb'});
     }
 
-    public static showGDPRPopup(nativeBridge: NativeBridge, configuration: Configuration, abGroup: number): Promise<boolean> {
-        if((abGroup === 18 || abGroup === 19) && this._euCountries.indexOf(configuration.getCountry()) !== -1) {
-            return nativeBridge.Storage.get(StorageType.PRIVATE, 'gdpr.popupshown.value').then(value => {
-                return !<boolean>value;
-            }).catch(([error]) => {
-                if (error === StorageError[StorageError.COULDNT_GET_VALUE]) {
-                    return true;
-                } else {
-                    return false;
-                }
-            });
-
-        } else {
-            return Promise.resolve(false);
-        }
-    }
-
-    private static _euCountries = [
-        'BE',
-        'BG',
-        'CZ',
-        'DK',
-        'DE',
-        'EE',
-        'IE',
-        'GR',
-        'ES',
-        'FR',
-        'HR',
-        'IT',
-        'CY',
-        'LV',
-        'LT',
-        'LU',
-        'HU',
-        'MT',
-        'NL',
-        'AT',
-        'PL',
-        'PT',
-        'RO',
-        'SI',
-        'SK',
-        'FI',
-        'SE',
-        'GB'
-    ];
 }