+++ src/ts/Views/EndScreen.ts
@@ -4,22 +4,20 @@ import SquareEndScreenTemplate from 'html/SquareEndScreen.html';
 import { NativeBridge } from 'Native/NativeBridge';
 import { View } from 'Views/View';
 import { Template } from 'Utilities/Template';
-import { IPrivacyHandler, Privacy } from 'Views/Privacy';
+import { IPrivacyHandler, AbstractPrivacy } from 'Views/AbstractPrivacy';
 import { Localization } from 'Utilities/Localization';
 import { AbstractAdUnit } from 'AdUnits/AbstractAdUnit';
 import { Campaign } from 'Models/Campaign';
 import { IEndScreenDownloadParameters } from 'EventHandlers/EndScreenEventHandler';
 import { AdUnitStyle } from 'Models/AdUnitStyle';
 import { CustomFeatures } from 'Utilities/CustomFeatures';
-import { Platform } from 'Constants/Platform';
 import { SquareEndScreenUtilities } from 'Utilities/SquareEndScreenUtilities';
 
 export interface IEndScreenHandler {
     onEndScreenDownload(parameters: IEndScreenDownloadParameters): void;
-    onEndScreenPrivacy(url: string): void;
     onEndScreenClose(): void;
     onKeyEvent(keyCode: number): void;
-    onOptOutPopupShown(clicked: boolean): void;
+    onGDPRPopupSkipped(): void;
 }
 
 const GDPR_OPT_OUT_BASE  = 'gdpr-pop-up-base';
@@ -30,24 +28,23 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
 
     protected _localization: Localization;
     protected _adUnitStyle?: AdUnitStyle;
-    private _coppaCompliant: boolean;
     private _gameName: string | undefined;
-    private _privacy: Privacy;
+    private _privacy: AbstractPrivacy;
     private _isSwipeToCloseEnabled: boolean = false;
     private _abGroup: number;
-    private _showOptOutPopup: boolean;
+    private _showGDPRBanner: boolean = false;
     private _gdprPopupClicked = false;
     private _campaignId: string | undefined;
     private _osVersion: string | undefined;
 
-    constructor(nativeBridge: NativeBridge, coppaCompliant: boolean, language: string, gameId: string, gameName: string | undefined, abGroup: number, adUnitStyle?: AdUnitStyle, showOptOutPopup: boolean = false, campaignId?: string, osVersion?: string) {
+    constructor(nativeBridge: NativeBridge, language: string, gameId: string, gameName: string | undefined, abGroup: number, privacy: AbstractPrivacy, showGDPRBanner: boolean, adUnitStyle?: AdUnitStyle, campaignId?: string, osVersion?: string) {
         super(nativeBridge, 'end-screen');
-        this._coppaCompliant = coppaCompliant;
         this._localization = new Localization(language, 'endscreen');
         this._abGroup = abGroup;
         this._gameName = gameName;
+        this._privacy = privacy;
         this._adUnitStyle = adUnitStyle;
-        this._showOptOutPopup = showOptOutPopup;
+        this._showGDPRBanner = showGDPRBanner;
         this._campaignId = campaignId;
         this._osVersion = osVersion;
 
@@ -71,7 +68,7 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
             },
             {
                 event: 'click',
-                listener: (event: Event) => this.onOptOutPopupEvent(event),
+                listener: (event: Event) => this.onGDPRPopupEvent(event),
                 selector: '.gdpr-link'
             }
         ];
@@ -85,6 +82,11 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
                 selector: '.campaign-container, .game-background, .btn.download'
             });
         }
+
+        this._privacy.render();
+        this._privacy.hide();
+        document.body.appendChild(this._privacy.container());
+        this._privacy.addEventHandler(this);
     }
 
     public render(): void {
@@ -130,35 +132,32 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
     public hide(): void {
         super.hide();
 
-        if (this._showOptOutPopup) {
-            this._handlers.forEach(handler => handler.onOptOutPopupShown(this._gdprPopupClicked));
+        if (this._showGDPRBanner && !this._gdprPopupClicked) {
+            this._handlers.forEach(handler => handler.onGDPRPopupSkipped());
         }
 
         if (this._privacy) {
             this._privacy.hide();
-            this._privacy.container().parentElement!.removeChild(this._privacy.container());
-            delete this._privacy;
         }
     }
 
     public onPrivacyClose(): void {
         if (this._privacy) {
-            this._privacy.removeEventHandler(this);
             this._privacy.hide();
-            this._privacy.container().parentElement!.removeChild(this._privacy.container());
-            delete this._privacy;
         }
     }
 
     public onPrivacy(url: string): void {
-        this._handlers.forEach(handler => handler.onEndScreenPrivacy(url));
+        // do nothing
+    }
+
+    public onGDPROptOut(optOutEnabled: boolean): void {
+        // do nothing
     }
 
     protected getEndscreenAlt(campaign?: Campaign) {
-        if (this._showOptOutPopup) {
-            if (this._abGroup === 18 || this._abGroup === 19) {
-                return GDPR_OPT_OUT_BASE;
-            }
+        if (this._showGDPRBanner) {
+            return GDPR_OPT_OUT_BASE;
         }
 
         const campaignId = campaign ? campaign.getId() : this._campaignId;
@@ -179,21 +178,15 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
 
     private onPrivacyEvent(event: Event): void {
         event.preventDefault();
-        this._privacy = new Privacy(this._nativeBridge, this._coppaCompliant);
-        this._privacy.render();
-        document.body.appendChild(this._privacy.container());
-        this._privacy.addEventHandler(this);
+
+        this._privacy.show();
     }
 
-    private onOptOutPopupEvent(event: Event) {
+    private onGDPRPopupEvent(event: Event) {
         event.preventDefault();
 
         this._gdprPopupClicked = true;
-
-        this._privacy = new Privacy(this._nativeBridge, this._coppaCompliant);
-        this._privacy.render();
-        document.body.appendChild(this._privacy.container());
-        this._privacy.addEventHandler(this);
+        this._privacy.show();
     }
 
     private getTemplate() {